requirements.txt
pan-os-python==1.11.0
pan-python==0.17.0
